feat: bundle the wasm binary up properly using parcel bundler and proper imports
