+++ .gitignore
@@ -1,2 +1,4 @@
 .DS_Store
 /public/
+/.parcel-cache/
+/node_modules/
