+++ .npmignore
@@ -1,5 +1,7 @@
 .DS_Store
 /public/
+/src/
+/.parcel-cache/
 /go.sum
 /go.mod
 *.go
